Update Weird Algorithm.cpp

+++ Weird Algorithm/Weird Algorithm.cpp
@@ -318,7 +318,7 @@ freopen("input.txt","r",stdin); //file input.txt is opened in reading mode i.e "
 freopen("output.txt","w",stdout);  //file output.txt is opened in writing mode i.e "w"
 #endif
         
-    int n;cin>>n;
+    long long n;cin>>n;
     while(n!=1)
     {
         cout<<n<<" ";
